rename widgets and ignore autolguon_dashboard module imports
